+++ x-pack/test/common/services/security_solution/roles/ess_roles.json
