+++ x-pack/test/detection_engine_api_integration/security_and_spaces/group10/create_signals_migrations.ts
@@ -12,7 +12,7 @@ import {
   DETECTION_ENGINE_SIGNALS_FINALIZE_MIGRATION_URL,
   DETECTION_ENGINE_SIGNALS_MIGRATION_URL,
 } from '@kbn/security-solution-plugin/common/constants';
-import { ROLES } from '@kbn/security-solution-plugin/common/test';
+import { ROLES } from 'x-pack/test/common/services/security_solution/roles';
 import { SIGNALS_TEMPLATE_VERSION } from '@kbn/security-solution-plugin/server/lib/detection_engine/routes/index/get_signals_template';
 import { Signal } from '@kbn/security-solution-plugin/server/lib/detection_engine/rule_types/types';
 import { FtrProviderContext } from '../../common/ftr_provider_context';
