+++ x-pack/test/detection_engine_api_integration/security_and_spaces/group10/read_privileges.ts
@@ -8,7 +8,7 @@
 import expect from '@kbn/expect';
 import { DETECTION_ENGINE_PRIVILEGES_URL } from '@kbn/security-solution-plugin/common/constants';
 
-import { ROLES } from '@kbn/security-solution-plugin/common/test';
+import { ROLES } from 'x-pack/test/common/services/security_solution/roles';
 import { FtrProviderContext } from '../../common/ftr_provider_context';
 import { createUserAndRole, deleteUserAndRole } from '../../../common/services/security_solution';
