+++ x-pack/test/security_solution_endpoint/apps/endpoint/responder.ts
@@ -8,7 +8,7 @@
 import { IndexedHostsAndAlertsResponse } from '@kbn/security-solution-plugin/common/endpoint/index_data';
 import { TimelineResponse } from '@kbn/security-solution-plugin/common/api/timeline';
 import { type IndexedEndpointRuleAlerts } from '@kbn/security-solution-plugin/common/endpoint/data_loaders/index_endpoint_rule_alerts';
-import { DATE_RANGE_OPTION_TO_TEST_SUBJ_MAP } from '@kbn/security-solution-plugin/common/test';
+import { DATE_RANGE_OPTION_TO_TEST_SUBJ_MAP } from 'x-pack/test/common/services/security_solution/roles';
 import { FtrProviderContext } from '../../ftr_provider_context';
 import { targetTags } from '../../target_tags';
