+++ x-pack/test/security_solution_ftr/page_objects/timeline/index.ts
@@ -6,7 +6,7 @@
  */
 
 import { subj as testSubjSelector } from '@kbn/test-subj-selector';
-import { DATE_RANGE_OPTION_TO_TEST_SUBJ_MAP } from '@kbn/security-solution-plugin/common/test';
+import { DATE_RANGE_OPTION_TO_TEST_SUBJ_MAP } from 'x-pack/test/common/services/security_solution/roles';
 import { FtrService } from '../../../functional/ftr_provider_context';
 
 const TIMELINE_BOTTOM_BAR_CONTAINER_TEST_SUBJ = 'flyoutBottomBar';
